:recycle: Factor out file deletion tests
